backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
Pillow>=10.0
django-cors-headers>=4.3
gunicorn>=21.0
psycopg2-binary>=2.9
dj-database-url>=2.1
whitenoise>=6.6
python-decouple>=3.8
cloudinary>=1.36
django-cloudinary-storage>=0.3.0
